requirements.txt
fastapi~=0.115.12
uvicorn
gunicorn
requests~=2.32.3
pydantic~=2.11.3
scikit-learn~=1.7.1
keybert~=0.9.0
python-dotenv